requirements.txt
python_dotenv==0.6.3
psycopg2==2.7.3.2
requests==2.18.4